+++ src/components/Top/OotdTop.tsx
@@ -1,5 +1,5 @@
-import { SimpleTopBar } from "./SimpleTopBar";
+import { SimpleTopBar, SimpleTopBarProps } from "./SimpleTopBar";
 
-export function OotdTop() {
-  return <SimpleTopBar title="OOTD" />;
+export function OotdTop({ onProfilePress }: SimpleTopBarProps) {
+  return <SimpleTopBar title="OOTD" onProfilePress={onProfilePress} />;
 }
