+++ src/components/Top/StatsTop.tsx
@@ -1,5 +1,5 @@
-import { SimpleTopBar } from "./SimpleTopBar";
+import { SimpleTopBar, SimpleTopBarProps } from "./SimpleTopBar";
 
-export function StatsTop() {
-  return <SimpleTopBar title="통계" />;
+export function StatsTop({ onProfilePress }: SimpleTopBarProps) {
+  return <SimpleTopBar title="통계" onProfilePress={onProfilePress} />;
 }
