+++ src/screens/Mypage/NicknameChangeScreen.tsx
@@ -8,7 +8,7 @@ import NicknameButton from "@/components/Buttons/nickname_button/NicknameButton"
 import { Colors } from "@/theme/colors";
 import Alert from "@/components/Alert/Alert";
 import type { NativeStackScreenProps } from "@react-navigation/native-stack";
-import { RootStackParamList } from "../../../App";
+import { RootStackParamList } from "@/screens/Test/TestPlace";
 
 type Props = NativeStackScreenProps<RootStackParamList, "NicknameEdit">;
